Introduce enum_utils for common code
