+++ plugins/module_utils/enum_utils.py
@@ -1,0 +1,39 @@
+#  Copyright: (c) 2024, Puzzle ITC
+#  GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
+"""Reusable Enum utilities"""
+from enum import Enum
+from typing import List
+
+
+class ListEnum(Enum):
+    """Enum class with some handy utility functions."""
+
+    @classmethod
+    def as_list(cls) -> List[str]:
+        """
+        Return a list
+        Returns
+        -------
+
+        """
+        return [entry.value for entry in cls]
+
+    @classmethod
+    def from_string(cls, value: str) -> "ListEnum":
+        """
+        Returns Enum value, from a given String.
+        If no enum value can be mapped to the input string,
+        ValueError is raised.
+        Parameters
+        ----------
+        value: `str`
+            String to be mapped to enum value
+
+        Returns
+        -------
+        Enum value
+        """
+        for _key, _value in cls.__members__.items():
+            if value in (_key, _value.value):
+                return _value
+        raise ValueError(f"'{cls.__name__}' enum not found for '{value}'")
