+++ plugins/module_utils/system_access_users_utils.py
@@ -27,7 +27,6 @@ Licensed under the GNU General Public License v3.0+
 
 
 from dataclasses import dataclass, asdict, fields
-from enum import Enum
 from typing import List, Optional
 import base64
 import os
@@ -42,6 +41,7 @@ from ansible_collections.puzzle.opnsense.plugins.module_utils import (
 from ansible_collections.puzzle.opnsense.plugins.module_utils.config_utils import (
     OPNsenseModuleConfig,
 )
+from ansible_collections.puzzle.opnsense.plugins.module_utils.enum_utils import ListEnum
 
 
 class OPNSenseGroupNotFoundError(Exception):
@@ -62,40 +62,7 @@ class OPNSenseCryptReturnError(Exception):
     """
 
 
-class ListEnum(Enum):
-    """Enum class with some handy utility functions."""
-
-    @classmethod
-    def as_list(cls) -> List[str]:
-        """
-        Return a list
-        Returns
-        -------
-
-        """
-        return [entry.value for entry in cls]
-
-    @classmethod
-    def from_string(cls, value: str) -> "ListEnum":
-        """
-        Returns Enum value, from a given String.
-        If no enum value can be mapped to the input string,
-        ValueError is raised.
-        Parameters
-        ----------
-        value: `str`
-            String to be mapped to enum value
-
-        Returns
-        -------
-        Enum value
-        """
-        for _key, _value in cls.__members__.items():
-            if value in (_key, _value.value):
-                return _value
-        raise ValueError(f"'{cls.__name__}' enum not found for '{value}'")
-
-
+# pylint: disable=too-few-public-methods
 class UserLoginShell(ListEnum):
     """Represents the user login shell."""
